Refactoring based on Nerd's feedback
